+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Adapt Data Contract Specification in version 0.9.1
 - Use `models` section for `diff`/`breaking`
 - Add `model` command
+- Let inline print to STDOUT instead of overwriting datacontract file
 
 # 0.4.0
 - Basic implementation of `test` command for Soda Core